Add 'fill' button and improve mouse position
This adds a button to allow the user to change the background colour
over to any of the existing colours. It also improves the
positioning of the mouse cursor, so that it better matches up with
where the user is painting. Finally, it adjusts the sample config
so that the screen is bigger and the user can enjoy the full potential
of the canvas.