Make all ui-element vertical positioning relative
This makes sure all the paint buttons and brush buttons are
positioned vertically relative to previous buttons. This means
elements will not overlap if more are added. It is now more
compact; a user can display buttons for up to 8 colours on
the screen at once. We live in the future.